fix: fixed update command

+++ build.sh
@@ -1,3 +1,7 @@
 cd mspaint-src
-rm -rf Distribution/
-lune run Build bundle input="default.project.json" minify=true output="Distribution/Script.luau" env-name="Script" darklua-config-path="Build/DarkLua.json" temp-dir-base="Distribution" verbose=true
+
+if [ -d "Distribution" ]; then
+    rm -rf Distribution/
+fi
+
+bash build.sh
